requirements.txt
flask
opencv-python
requests
google.generativeai
python-dotenv
Pillow

